+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Added:
 * Added a "Start after Parent Session has completed" option to the Race Weekend Schedule popup
 * Added configurable negative points modifiers for championships for crashes and cuts.
 * Added an Entrant Attendance field to Championship table to help admins keep track of who is showing up for races.
+* Added "Register with Steam" to Championships with a Sign Up Form.
+* Added "Sign in with Steam" to the Update Account Details page.
 
 Fixes: